+++ core/src/main/java/uk/ac/bbsrc/tgac/miso/core/service/SampleService.java
@@ -1,13 +1,12 @@
 package uk.ac.bbsrc.tgac.miso.core.service;
 
 import java.io.IOException;
-import java.util.Collection;
 import java.util.List;
 
 import uk.ac.bbsrc.tgac.miso.core.data.Barcodable.EntityType;
 import uk.ac.bbsrc.tgac.miso.core.data.Sample;
-import uk.ac.bbsrc.tgac.miso.core.data.SampleIdentity;
 import uk.ac.bbsrc.tgac.miso.core.data.impl.view.EntityReference;
+import uk.ac.bbsrc.tgac.miso.core.data.impl.view.IdentityView;
 import uk.ac.bbsrc.tgac.miso.core.util.PaginatedDataSource;
 
 public interface SampleService
@@ -20,7 +19,7 @@ public interface SampleService
 
   public List<Sample> list() throws IOException;
 
-  public Collection<SampleIdentity> getIdentitiesByExternalNameOrAliasAndProject(String externalName, Long projectId, boolean exactMatch)
+  public List<IdentityView> getIdentitiesByExternalNameOrAliasAndProject(String externalName, Long projectId, boolean exactMatch)
       throws IOException;
 
   public void confirmExternalNameUniqueForProjectIfRequired(String externalNames, Sample sample) throws IOException;